Project upload
This Arduino library implements a very powerful class of timer.  By
default up to five timers can execute simultaneously; however, this
limit can be changed by modifying a single #define.  When each timer
expires its callback function is executed and is passed the pointer to a
user-defined object.   This pointer can be used in the callback function
to manipulate the object.

As written this library utilizes Timer/Counter 2 to interrupt the
processor on overflow.  The interrupt service routine increments a 15
bit clock and checks whether the new time equals any of the timers, each
timer that matches has its callback executed with its object pointer as
argument, the callback manipulates the object, a repeat count is
decremented and if the new count is zero the timer is removed from the
list.  If the repeat count is set to zero outside the service routine,
the timer repeats infinitely.

Two examples are provided to help new users get started.